--- UwU/Assets/ScoreKeeper.cs
@@ -1,25 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ScoreKeeper : MonoBehaviour
-{
-public int pointsPerPickup; 
-private int currentScore; 
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public int IncreaseScore(){
-        return (currentScore += pointsPerPickup); 
-    }
-}
