+++ UwU/Assets/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cee2cc66ae8ec534cbbd75b6442fbae3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
