+++ UwU/Assets/Script/ScoreKeeper.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ScoreKeeper : MonoBehaviour
+{
+public int pointsPerPickup; 
+private int currentScore; 
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        currentScore = 0; 
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void IncreaseScore(){
+        currentScore += pointsPerPickup; 
+    }
+
+    public int CurrentScore
+    {
+        get 
+        {
+            return currentScore; 
+        }
+    }
+}
