+++ UwU/Assets/Script/ScoreKeeper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a3fed812afc263143b2bd3ff0de8eb95
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
