+++ UwU/Library/LastSceneManagerSetup.txt
@@ -1,1 +1,5 @@
-sceneSetups: []
+sceneSetups:
+- path: Assets/Scenes/Week9&10.unity
+  isLoaded: 1
+  isActive: 1
+  isSubScene: 0
